Third attempt at JudgingCest
Last time I tried this I forgot that a team can only have one enqueued solution at a time.
So in this rendition, I create and then delete a bunch of teams to submit things at the same time.
It's untested but hopefully not fundamentally flawed like the last two were.