Added execution and judging code to Submission
Remember when I said Submission was close to an MVP state?
Well I still think that but last time I said that I was about 100 lines short, so who knows how close it is
I also added documentation to some of the functions.
None of this code is tested btw. Hoping to do that soon