Added features suggested by Geisler
The biggest of these features is running the gcc preprocessor before checking for forbidden words
I also added the ability to ignore the stderr output of compiled solutions.
Diffs are now now checked with the exit code of the command-line tool instead of checking the size of the output file.
Tested and working