Submissions now each have their own directory
So each solution submitted to touche gets it's own directory in the queue or judged folder.
This directory contains the source code initially, but gets used to store all the other files associated with the submission as well.
This required several small changes across a variety of files but it is
Tested and working