Refactor of submission.py in preparation for changes
Previously, I intended the Submission class to expose a series of methods to the user which would be called one after the other.
I've decided to go with a simpler API which requires only one method call.
Instead of thowing exceptions, I'm going to return a results object from that one method call.
In this commit I mainly rearranged the order of the methods in the class to make more sense for the new API