Reworked the way that responses are reported in the cronscript code
I created a SubmissionResults class that stores the results for a sumbission.
It has getter and setter methods to make sure data are stored in the correct format.

The Submission class now returns one of these objects with the new function get_results,
and then the cronscript sends those results off to the DBDriver.

The DBDriver class now reports results properly.
Before it was setting a column in the database which is meant to be set by a human judge only.

CSubmission and JavaSubmission were updated to work with the SubmissionResults class.

The exceptions are no longer used and were therefore removed.