Changed the AUTO_RESPONSES and RESPONSES tables and updated pages
The old structure of these tables was lacking-a-primary-key bad, and the new cronscript puts more information in the database.
So I changed dbcreate.sql accordingly.
I also changed the ordering of a few error codes based on conversations I had with Geisler.

These changes eliminated the need for the duplicated information found in responses.inc so that went bye-bye.

I heavily refactored judge/judge_response.php and fixed/tweaked several things in the UI.
The UI should look largely the same, but it now displays runtime error text and has a few less bugs.

I made some tweaks to JudgingCest because somebody hardcoded a value in there.
I also changed some text in JudgingCest to reflect a change in the UI

The other files in this commit have small refactoring changes because of the changes made to the database.

Tested and working