Added chroot functionality to the new cronscript
In chroot_wrapper.c, Geisler and I added the ability to kill it's child process when it recieves a SIGTERM signal.
This solves a problem where the cronscript would kill the chroot_wrapper parent process but not the child process when the time limit was exceeded.

Other changes include implementing the move_to and from_jail methods in Submission
and adding functionality in Submission's execute method to interface with chroot_wrapper properly.