The new cronscript now supports Python
There is also code in here to support the EMAXOUTPUT response, but that does not work exactly as it should.
Currently, the output is being limited and an error is being reported, but it is the wrong error in some cases.

JudgingCest also tests Python now, and those tests pass.