The contest now calls the cronscript to preform a test compile
The cronscript code now supports this feature.
This entailed relatively small changes to db_driver, submission, and cronscript

We decided to have the PHP and cronscript communicate much like they do in other cases.
This means that it should be easy to have successful compiles automatically submitted or something similar.
The JUDGED_SUBMISSIONS and QUEUED_SUBMISSIONS tables now have a new boolean TEST_COMPILE column.

This means that other files that used those tables had to be tweaked to filter out the TEST_COMPILE submissions.